+++ CHANGELOG.md
@@ -2,17 +2,25 @@
 
 ## CHANGELOG
 
-### 3.0.0 (26-Jul-2025)
-
+### 3.0.0 (19-Aug-2025)
 - First (attempt at a) MDM-agnostic release
 
----
+### 2.3.0 (19-Aug-2025)
+- Enhanced `operationMode` to verbosely execute when set to `debug` (Addresses Issue #28)
+- Adjusted GlobalProtect VPN check for IPv6
+- Enhanced `checkJssCertificateExpiration` function (Addresses Issue #27 via Pull Request #30; thanks, @theahadub and @ScottEKendall)
+- Extended Network Checks (Pull Request #31 addresses Issue #23; thanks big bunches, @tonyyo11!)
+- Added `organizationBrandingBannerURL` (thanks for the inspiration, @ScottEKendall!) [Image by benzoix on Freepik](https://www.freepik.com/author/benzoix)
 
-### 2.2.0 (25-Jul-2025)
+### 2.2.0 (15-Aug-2025)
 - Improved the GlobalProtect VPN IP detection logic
 - Added an option to show if an app is installed (Feature Request #18; thanks, @ScottEKendall!)
 - Add framework for different VPN clients and an internal VPN Client Check (Pull Request #16; thanks for another one, @HowardGMac!)
 - Addressed MHC does not show SF Symbols in the upper left corner - needs region check (Issue #21; thanks, @hbokh!)
+- Active IP Address section changes (Pull Request #24; thanks, Obi-@HowardGMac!)
+- Use zsh expansion in the `checkExternal` function to convert the results to lowercase so that the user doesn't have to match the case exactly in their results (Pull Request #25; thanks, @ScottEKendall!)
+- Added Tailscale VPN check (thanks, @alexfinn!)
+- Change zsh logic flag for Dialog check / installation from `-e` to `-x` to make sure file exists and is executable (Pull Request #26; thanks, @ScottEKendall!)
 
 ### 2.1.0 (24-Jul-2025)
 - Added an `operationMode` of "debug" to specifically enable swiftDialog debugging